requirements.txt
django
djangorestframework
djangorestframework-simplejwt
djangorestframework-jwt
django-cors-headers
django-jsonfield